refactor(commands): extract user tests to sibling file
